% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figr8.R
\name{get_page}
\alias{get_page}
\title{Get a single page from the Figure 8 API}
\usage{
get_page(job_id, endpoint, ...)
}
\arguments{
\item{job_id}{the Job ID}

\item{endpoint}{the name of the page without the .json}
}
\description{
Get a single page from the Figure 8 API
}
