% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figr8.R
\name{results}
\alias{results}
\title{Retrieve job results}
\usage{
results(job_id)
}
\arguments{
\item{job_id}{the Job ID}

\item{metacols}{a vector of column names to retrieve as metadata (i.e., one per row)}

\item{targetcol}{the column to retrieve as coded value (i.e. one per judgment)}
}
\description{
Retrieve job results
}
